Refactoring Not Valid Field Class to DTOs
